feat(gameserver): separate transport layer — Transport/* with full impl, netcode re-export, no dummy
